requirements.txt
Django==5.1         # Versión de Django
djangorestframework==3.15.2   # Django REST Framework
mysqlclient==2.2.4
django-cors-headers==4.4.0  # Asegúrate de usar la versión correcta instalada
djangorestframework-simplejwt==5.3.1
django-extensions==3.2.3
google-generativeai==0.8.3

